add SwarmHandler and MixnetHandler
